+++ game/build.gradle
@@ -27,21 +27,6 @@ dependencies {
     testCompile group: 'org.springframework.boot', name: 'spring-boot-starter-test', version: '2.0.0.RELEASE'
 }
 
-/*springBoot {
-    mainClass = "mm.MatchMakerApp"
-}
-sourceSets {
-    main {
-        java {
-            srcDirs = ['src/main/java']
-        }
-    }
-    test {
-        java {
-            srcDirs = ['src/test/java']
-        }
-    }
-}*/
 checkstyle {
     ignoreFailures = false
     toolVersion = '7.5'
