+++ game/src/main/java/mm/Repo/GameSessionsRepository.java
@@ -1,7 +1,5 @@
-package mm;
+package mm.Repo;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.context.annotation.Bean;
 import org.springframework.stereotype.Repository;
