+++ game/src/main/java/playerdb/TestRegistration.java
@@ -1,4 +1,4 @@
-package mm;
+package playerdb;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
